+++ extensions/yii/elasticsearch/ActiveRecord.php
@@ -44,9 +44,8 @@ use yii\helpers\StringHelper;
  */
 class ActiveRecord extends BaseActiveRecord
 {
-	const PRIMARY_KEY_NAME = 'id';
-
 	private $_id;
+	private $_score;
 	private $_version;
 
 	/**
@@ -67,14 +66,6 @@ class ActiveRecord extends BaseActiveRecord
 	{
 		$query = static::createQuery();
 		if (is_array($q)) {
-			if (count($q) == 1 && (array_key_exists(ActiveRecord::PRIMARY_KEY_NAME, $q)) && $query->where === null) {
-				$pk = $q[ActiveRecord::PRIMARY_KEY_NAME];
-				if (is_array($pk)) {
-					return  static::mget($pk);
-				} else {
-					return static::get($pk);
-				}
-			}
 			return $query->andWhere($q)->one();
 		} elseif ($q !== null) {
 			return static::get($q);
@@ -155,9 +146,12 @@ class ActiveRecord extends BaseActiveRecord
 
 	// TODO implement copy and move as pk change is not possible
 
-	public function getId()
+	/**
+	 * @return float returns the score of this record when it was retrieved via a [[find()]] query.
+	 */
+	public function getScore()
 	{
-		return $this->_id;
+		return $this->_score;
 	}
 
 	/**
@@ -165,10 +159,11 @@ class ActiveRecord extends BaseActiveRecord
 	 * @param mixed $value
 	 * @throws \yii\base\InvalidCallException when record is not new
 	 */
-	public function setId($value)
+	public function setPrimaryKey($value)
 	{
-		if ($this->isNewRecord) {
-			$this->_id = $value;
+		$pk = static::primaryKey()[0];
+		if ($this->getIsNewRecord() || $pk != '_id') {
+			$this->$pk = $value;
 		} else {
 			throw new InvalidCallException('Changing the primaryKey of an already saved record is not allowed.');
 		}
@@ -179,10 +174,11 @@ class ActiveRecord extends BaseActiveRecord
 	 */
 	public function getPrimaryKey($asArray = false)
 	{
+		$pk = static::primaryKey()[0];
 		if ($asArray) {
-			return [ActiveRecord::PRIMARY_KEY_NAME => $this->_id];
+			return [$pk => $this->$pk];
 		} else {
-			return $this->_id;
+			return $this->$pk;
 		}
 	}
 
@@ -191,30 +187,53 @@ class ActiveRecord extends BaseActiveRecord
 	 */
 	public function getOldPrimaryKey($asArray = false)
 	{
-		$id = $this->isNewRecord ? null : $this->_id;
-		if ($asArray) {
-			return [ActiveRecord::PRIMARY_KEY_NAME => $id];
+		$pk = static::primaryKey()[0];
+		if ($this->getIsNewRecord()) {
+			$id = null;
+		} elseif ($pk == '_id') {
+			$id = $this->_id;
 		} else {
-			return $this->_id;
+			$id = $this->getOldAttribute($pk);
+		}
+		if ($asArray) {
+			return [$pk => $id];
+		} else {
+			return $id;
 		}
 	}
 
 	/**
-	 * This method defines the primary.
+	 * This method defines the attribute that uniquely identifies a record.
 	 *
-	 * The primaryKey for elasticsearch documents is always `primaryKey`. It can not be changed.
+	 * The primaryKey for elasticsearch documents is the `_id` field by default. This field is not part of the
+	 * ActiveRecord attributes so you should never add `_id` to the list of [[attributes()|attributes]].
 	 *
-	 * @return string[] the primary keys of this record.
+	 * You may overide this method to define the primary key name when you have defined
+	 * [path mapping](http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/mapping-id-field.html)
+	 * for the `_id` field so that it is part of the `_source` and thus part of the [[attributes()|attributes]].
+	 *
+	 * Note that elasticsearch only supports _one_ attribute to be the primary key. However to match the signature
+	 * of the [[\yii\db\ActiveRecordInterface|ActiveRecordInterface]] this methods returns an array instead of a
+	 * single string.
+	 *
+	 * @return string[] array of primary key attributes. Only the first element of the array will be used.
 	 */
 	public static function primaryKey()
 	{
-		return [ActiveRecord::PRIMARY_KEY_NAME];
+		return ['_id'];
 	}
 
 	/**
 	 * Returns the list of all attribute names of the model.
+	 *
 	 * This method must be overridden by child classes to define available attributes.
-	 * @return array list of attribute names.
+	 *
+	 * Attributes are names of fields of the corresponding elasticsearch document.
+	 * The primaryKey for elasticsearch documents is the `_id` field by default which is not part of the attributes.
+	 * You may define [path mapping](http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/mapping-id-field.html)
+	 * for the `_id` field so that it is part of the `_source` fields and thus becomes part of the attributes.
+	 *
+	 * @return string[] list of attribute names.
 	 */
 	public function attributes()
 	{
@@ -246,8 +265,11 @@ class ActiveRecord extends BaseActiveRecord
 	 */
 	public static function create($row)
 	{
-		$row['_source'][ActiveRecord::PRIMARY_KEY_NAME] = $row['_id'];
 		$record = parent::create($row['_source']);
+		$pk = static::primaryKey()[0];
+		$record->$pk = $row['_id'];
+		$record->_score = isset($row['_score']) ? $row['_score'] : null;
+		$record->_version = isset($row['_version']) ? $row['_version'] : null; // TODO version should always be available...
 		return $record;
 	}
 
@@ -317,11 +339,16 @@ class ActiveRecord extends BaseActiveRecord
 				$options
 			);
 
-			if (!$response['ok']) {
+			if (!isset($response['ok'])) {
 				return false;
 			}
-			$this->_id = $response['_id'];
+			$pk = static::primaryKey()[0];
+			$this->$pk = $response['_id'];
+			if ($pk != '_id') {
+				$values[$pk] = $response['_id'];
+			}
 			$this->_version = $response['_version'];
+			$this->_score = null;
 			$this->setOldAttributes($values);
 			$this->afterSave(true);
 			return true;
@@ -344,16 +371,17 @@ class ActiveRecord extends BaseActiveRecord
 	 */
 	public static function updateAll($attributes, $condition = [])
 	{
-		if (count($condition) == 1 && isset($condition[ActiveRecord::PRIMARY_KEY_NAME])) {
-			$primaryKeys = (array) $condition[ActiveRecord::PRIMARY_KEY_NAME];
+		$pkName = static::primaryKey()[0];
+		if (count($condition) == 1 && isset($condition[$pkName])) {
+			$primaryKeys = is_array($condition[$pkName]) ? $condition[$pkName] : [$condition[$pkName]];
 		} else {
-			$primaryKeys = static::find()->where($condition)->column(ActiveRecord::PRIMARY_KEY_NAME);
+			$primaryKeys = static::find()->where($condition)->column($pkName); // TODO check whether this works with default pk _id
 		}
 		if (empty($primaryKeys)) {
 			return 0;
 		}
 		$bulk = '';
-		foreach((array) $primaryKeys as $pk) {
+		foreach($primaryKeys as $pk) {
 			$action = Json::encode([
 				"update" => [
 					"_id" => $pk,
@@ -371,10 +399,16 @@ class ActiveRecord extends BaseActiveRecord
 		$url = [static::index(), static::type(), '_bulk'];
 		$response = static::getDb()->post($url, [], $bulk);
 		$n=0;
+		$errors = [];
 		foreach($response['items'] as $item) {
-			if ($item['update']['ok']) {
+			if (isset($item['update']['error'])) {
+				$errors[] = $item['update'];
+			} elseif ($item['update']['ok']) {
 				$n++;
 			}
+		}
+		if (!empty($errors)) {
+			throw new Exception(__METHOD__ . ' failed updating records.', $errors);
 		}
 		return $n;
 	}
@@ -395,16 +429,17 @@ class ActiveRecord extends BaseActiveRecord
 	 */
 	public static function updateAllCounters($counters, $condition = [])
 	{
-		if (count($condition) == 1 && isset($condition[ActiveRecord::PRIMARY_KEY_NAME])) {
-			$primaryKeys = (array) $condition[ActiveRecord::PRIMARY_KEY_NAME];
+		$pkName = static::primaryKey()[0];
+		if (count($condition) == 1 && isset($condition[$pkName])) {
+			$primaryKeys = is_array($condition[$pkName]) ? $condition[$pkName] : [$condition[$pkName]];
 		} else {
-			$primaryKeys = static::find()->where($condition)->column(ActiveRecord::PRIMARY_KEY_NAME);
+			$primaryKeys = static::find()->where($condition)->column($pkName); // TODO check whether this works with default pk _id
 		}
 		if (empty($primaryKeys) || empty($counters)) {
 			return 0;
 		}
 		$bulk = '';
-		foreach((array) $primaryKeys as $pk) {
+		foreach($primaryKeys as $pk) {
 			$action = Json::encode([
 				"update" => [
 					"_id" => $pk,
@@ -426,12 +461,17 @@ class ActiveRecord extends BaseActiveRecord
 		// TODO do this via command
 		$url = [static::index(), static::type(), '_bulk'];
 		$response = static::getDb()->post($url, [], $bulk);
-
 		$n=0;
+		$errors = [];
 		foreach($response['items'] as $item) {
-			if ($item['update']['ok']) {
+			if (isset($item['update']['error'])) {
+				$errors[] = $item['update'];
+			} elseif ($item['update']['ok']) {
 				$n++;
 			}
+		}
+		if (!empty($errors)) {
+			throw new Exception(__METHOD__ . ' failed updating records counters.', $errors);
 		}
 		return $n;
 	}
@@ -452,16 +492,17 @@ class ActiveRecord extends BaseActiveRecord
 	 */
 	public static function deleteAll($condition = [])
 	{
-		if (count($condition) == 1 && isset($condition[ActiveRecord::PRIMARY_KEY_NAME])) {
-			$primaryKeys = (array) $condition[ActiveRecord::PRIMARY_KEY_NAME];
+		$pkName = static::primaryKey()[0];
+		if (count($condition) == 1 && isset($condition[$pkName])) {
+			$primaryKeys = is_array($condition[$pkName]) ? $condition[$pkName] : [$condition[$pkName]];
 		} else {
-			$primaryKeys = static::find()->where($condition)->column(ActiveRecord::PRIMARY_KEY_NAME);
+			$primaryKeys = static::find()->where($condition)->column($pkName); // TODO check whether this works with default pk _id
 		}
 		if (empty($primaryKeys)) {
 			return 0;
 		}
 		$bulk = '';
-		foreach((array) $primaryKeys as $pk) {
+		foreach($primaryKeys as $pk) {
 			$bulk .= Json::encode([
 				"delete" => [
 					"_id" => $pk,
@@ -475,10 +516,16 @@ class ActiveRecord extends BaseActiveRecord
 		$url = [static::index(), static::type(), '_bulk'];
 		$response = static::getDb()->post($url, [], $bulk);
 		$n=0;
+		$errors = [];
 		foreach($response['items'] as $item) {
-			if ($item['delete']['found'] && $item['delete']['ok']) {
+			if (isset($item['delete']['error'])) {
+				$errors[] = $item['delete'];
+			} elseif ($item['delete']['found'] && $item['delete']['ok']) {
 				$n++;
 			}
+		}
+		if (!empty($errors)) {
+			throw new Exception(__METHOD__ . ' failed deleting records.', $errors);
 		}
 		return $n;
 	}