+++ extensions/yii/elasticsearch/CHANGELOG.md
@@ -5,6 +5,8 @@ Yii Framework 2 elasticsearch extension Change Log
 ----------------------------
 
 - Enh #1382: Added a debug toolbar panel for elasticsearch (cebe)
+- Enh #1765: Added support for primary key path mapping, pk can now be part of the attributes when mapping is defined (cebe)
+- Chg #1765: Changed handling of ActiveRecord primary keys, removed getId(), use getPrimaryKey() instead (cebe)
 
 2.0.0 alpha, December 1, 2013
 -----------------------------